requirements.txt

Flask
pandas
python-dotenv
gunicorn
twilio
